another major refactor bruh
