+++ crates/sozo/src/commands/build.rs
@@ -1,13 +1,14 @@
 use anyhow::Result;
 use clap::Args;
-use scarb::core::Workspace;
+use scarb::core::Config;
 use scarb::ops;
 
 #[derive(Args, Debug)]
 pub struct BuildArgs;
 
 impl BuildArgs {
-    pub fn run(self, ws: &Workspace<'_>) -> Result<()> {
-        ops::compile(ws)
+    pub fn run(self, config: &Config) -> Result<()> {
+        let ws = scarb::ops::read_workspace(config.manifest_path(), config)?;
+        ops::compile(&ws)
     }
 }
