+++ crates/sozo/src/commands/migrate.rs
@@ -1,12 +1,11 @@
 use anyhow::{Context, Result};
 use clap::Args;
-use scarb::core::Workspace;
+use scarb::core::Config;
 
 use super::options::account::AccountOptions;
 use super::options::dojo_metadata_from_workspace;
 use super::options::starknet::StarknetOptions;
 use super::options::world::WorldOptions;
-use crate::commands::build::BuildArgs;
 use crate::ops::migration;
 
 #[derive(Args)]
@@ -26,15 +25,17 @@ pub struct MigrateArgs {
 }
 
 impl MigrateArgs {
-    pub fn run(self, ws: &Workspace<'_>) -> Result<()> {
+    pub fn run(self, config: &Config) -> Result<()> {
+        let ws = scarb::ops::read_workspace(config.manifest_path(), config)?;
+
         let target_dir = ws.target_dir().path_existent().unwrap();
         let target_dir = target_dir.join(ws.config().profile().as_str());
 
         if !target_dir.join("manifest.json").exists() {
-            BuildArgs.run(ws)?;
+            scarb::ops::compile(&ws)?;
         }
 
-        let mut env_metadata = dojo_metadata_from_workspace(ws)
+        let mut env_metadata = dojo_metadata_from_workspace(&ws)
             .and_then(|dojo_metadata| dojo_metadata.get("env").cloned());
 
         // If there is an environment-specific metadata, use that, otherwise use the
@@ -45,7 +46,7 @@ impl MigrateArgs {
             .or(env_metadata);
 
         ws.config().tokio_handle().block_on(async {
-            let world_address = self.world.address(ws).ok();
+            let world_address = self.world.address(&ws).ok();
             let provider = self.starknet.provider(env_metadata.as_ref())?;
 
             let account = self
