+++ crates/sozo/src/commands/mod.rs
@@ -1,16 +1,7 @@
 use anyhow::Result;
-use camino::Utf8PathBuf;
-use clap::{Parser, Subcommand};
-use scarb::compiler::Profile;
-use scarb::ui;
-use smol_str::SmolStr;
-use tracing::level_filters::LevelFilter;
-use tracing_log::AsTrace;
+use scarb::core::Config;
 
-use self::build::BuildArgs;
-use self::init::InitArgs;
-use self::migrate::MigrateArgs;
-use self::test::TestArgs;
+use crate::args::Commands;
 
 pub(crate) mod build;
 pub(crate) mod init;
@@ -18,78 +9,11 @@ pub(crate) mod migrate;
 pub(crate) mod options;
 pub(crate) mod test;
 
-#[derive(Subcommand)]
-pub enum Commands {
-    #[command(about = "Build the world, generating the necessary artifacts for deployment")]
-    Build(BuildArgs),
-    #[command(about = "Initialize a new project")]
-    Init(InitArgs),
-    #[command(about = "Run a migration, declaring and deploying contracts as necessary to \
-                       update the world")]
-    Migrate(Box<MigrateArgs>),
-    #[command(about = "Test the project's smart contracts")]
-    Test(TestArgs),
-}
-
-#[derive(Parser)]
-#[command(author, version, about, long_about = None)]
-#[command(propagate_version = true)]
-pub struct App {
-    #[arg(long)]
-    #[arg(hide_short_help = true)]
-    #[arg(env = "DOJO_MANIFEST_PATH")]
-    #[arg(help = "Override path to a directory containing a Scarb.toml file.")]
-    pub manifest_path: Option<Utf8PathBuf>,
-
-    #[clap(help = "Specify the profile to use.")]
-    #[command(flatten)]
-    pub profile_spec: ProfileSpec,
-
-    #[clap(help = "Logging verbosity.")]
-    #[command(flatten)]
-    pub verbose: clap_verbosity_flag::Verbosity,
-
-    #[command(subcommand)]
-    pub command: Commands,
-}
-
-impl App {
-    pub fn ui_verbosity(&self) -> ui::Verbosity {
-        let filter = self.verbose.log_level_filter().as_trace();
-        if filter >= LevelFilter::WARN {
-            ui::Verbosity::Verbose
-        } else if filter > LevelFilter::OFF {
-            ui::Verbosity::Normal
-        } else {
-            ui::Verbosity::Quiet
-        }
-    }
-}
-
-/// Profile specifier.
-#[derive(Parser, Clone, Debug)]
-#[group(multiple = false)]
-pub struct ProfileSpec {
-    #[arg(short = 'P', long)]
-    #[arg(help = "Specify profile to use by name.")]
-    pub profile: Option<SmolStr>,
-
-    #[arg(long, hide_short_help = true)]
-    #[arg(help = "Use release profile.")]
-    pub release: bool,
-
-    #[arg(long, hide_short_help = true)]
-    #[arg(help = "Use dev profile.")]
-    pub dev: bool,
-}
-
-impl ProfileSpec {
-    pub fn determine(&self) -> Result<Profile> {
-        Ok(match &self {
-            Self { release: true, .. } => Profile::RELEASE,
-            Self { dev: true, .. } => Profile::DEV,
-            Self { profile: Some(profile), .. } => Profile::new(profile.clone())?,
-            _ => Profile::default(),
-        })
+pub fn run(command: Commands, config: &Config) -> Result<()> {
+    match command {
+        Commands::Init(args) => args.run(),
+        Commands::Test(args) => args.run(config),
+        Commands::Build(args) => args.run(config),
+        Commands::Migrate(args) => args.run(config),
     }
 }
